Improve button texture
Texture taken from https://github.com/kyrptonaught/Inventory-Sorter/pull/108,
credit to https://github.com/Abbie5

Unfortunately I was unable to make the rest of the
changes in that MR work with the 1.21 codebase, so
the button still looks pretty wonky. I still plan
removing the buttons in an upcoming version, but
for now I consider this an improvement.